comments/serializers.py
from rest_framework import serializers


from comments.models import Comment


class CommentSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Comment
        fields = ['content', 'post']