pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["hrr2000 <thedragonea@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.3"
django-model-utils = "^4.3.1"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^3.7.0"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

